+++ app/Models/Comment.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Comment extends Model
+{
+    protected $fillable = [
+        'created_by',
+        'comment',
+        'created_at',
+        'updated_at',
+        'contact_id'
+    ];
+
+    public function contact()
+    {
+        return $this->belongsTo(Contact::class);
+    }
+}