+++ resources/views/contact/show.blade.php
@@ -30,5 +30,15 @@
                 </div>
             </div>
         </div>
+
+        <div class="row justify-content-center">
+            <div class="col-md-12">
+
+                @include('partials.comment.index')
+
+                @include('partials.comment.create')
+
+            </div>
+        </div>
     </div>
 @endsection